Use a public context baggage key
In order to facilitate the deprecation of this library and allow a
smoother migration to other logging libraries, this library will now
store the baggage values in a public key string, which can be read by
anyone, so even if someone sets the key by this deprecated library, we
still can read it by a newer library in another layer.

This is backwards compatible, since we're just making private things
public, except for the unlikely case where the key would collide with
someone other's key.